% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_4_Clonal_exploration.R
\name{calculate_clone}
\alias{calculate_clone}
\title{4_Clonal_exploration}
\usage{
calculate_clone(
  seq_df,
  clonotype,
  AA_or_NT = "NT",
  region = "CDR3",
  percentage = 100,
  calculate_shared_clones
)
}
\value{
The return value, if any, from executing the utility.
}
\description{
A utils function
}
\keyword{internal}
