% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_5_Feature_exploration.R
\name{draw_feature_violinplot}
\alias{draw_feature_violinplot}
\title{5_Feature_exploration}
\usage{
draw_feature_violinplot(
  values,
  name_values,
  sequence_info_df,
  group_info,
  additional_group_info = "Patient",
  show_reconstructed,
  compare_opposites = F,
  selected_rows,
  selected_subclones = NULL,
  selected_clones = NULL,
  hide_dots = FALSE,
  really_hide_dots = F,
  width = 1400,
  height = 1000,
  img_type = "png",
  scale = 4,
  seed = 1234,
  source = "feature_violinplot",
  primary_color = "#09BC8A"
)
}
\value{
The return value, if any, from executing the utility.
}
\description{
A utils function
}
\keyword{internal}
