% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_4_Clonal_exploration.R
\name{draw_sharedclonesplot}
\alias{draw_sharedclonesplot}
\title{4_Clonal_exploration}
\usage{
draw_sharedclonesplot(
  seq_df,
  sets = NULL,
  group = "Patient",
  selected_rows,
  clonotype,
  AA_or_NT = "AA",
  region = "CDR3",
  percentage = 100,
  freq_filter = 0,
  Selected_clones = NULL,
  dominance_threshold,
  color_by = "Clonal sharing",
  colorblind,
  Big_mem_color_values,
  seed = 1234,
  width = 1400,
  height = 1000,
  img_type = "png",
  scale = 4
)
}
\value{
The return value, if any, from executing the utility.
}
\description{
A utils function
}
\keyword{internal}
