% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the AbSolution Shiny Application
Launches the interactive AbSolution Shiny application for exploring
B-cell and T-cell receptor repertoire data. The application provides
tools for loading AIRR-formatted datasets, inspecting sequence
characteristics, visualizing clonotype distributions, and generating
summary reports.}
\usage{
run_app(
  verbose = FALSE,
  onStart = NULL,
  options = list(),
  enableBookmarking = "server",
  uiPattern = "/",
  ...
)
}
\arguments{
\item{verbose}{Logical. If \code{FALSE} (default), suppresses all warnings
and messages during app execution. Set to \code{TRUE} to see full console
output for debugging.}

\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
Internally, `run_app()` calls [`shiny::shinyApp()`] to start the
application, using the UI defined in [\code{app_ui()}] and the server
logic defined in [\code{app_server()}].
}
\details{
This function is the entry point that end-users should call after
installing the package.
}
\examples{
if (interactive()) {
  ### Launch the full application with default settings:
  run_app()
}
}
