% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{validate_before_run}
\alias{validate_before_run}
\title{Check parameters in 3. are properly selected}
\usage{
validate_before_run(
  work_as_categories,
  use_sharedVDJ,
  VJ_included,
  VDJ_maximize_clones,
  clonal_group
)
}
\description{
Check parameters in 3. are properly selected
}
\keyword{internal}
