% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_Data_processing.R
\name{big_PCA}
\alias{big_PCA}
\title{Big PCA}
\usage{
big_PCA(FBM, rows, columns)
}
\arguments{
\item{FBM}{A FBM object.}

\item{rows}{Index of rows of the FBM that will be used to calculate the PCA.}

\item{columns}{Index of columns of the FBM that will be used to calculate the PCA..}
}
\value{
A list with the PCA scores and its explained variances
}
\description{
A function to perform PCA over a FBM object.
}
\keyword{internal}
