% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_4_Clonal_exploration.R
\name{draw_upsetplot}
\alias{draw_upsetplot}
\title{4_Clonal_exploration}
\usage{
draw_upsetplot(
  seq_df,
  group = "Patient",
  selected_rows,
  clonotype,
  AA_or_NT = "AA",
  region = "CDR3",
  percentage = 100,
  freq_filter = 0,
  Selected_clones = NULL
)
}
\value{
The return value, if any, from executing the utility.
}
\description{
A utils function
}
\keyword{internal}
