% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_4_Clonal_exploration.R
\name{draw_violinplots}
\alias{draw_violinplots}
\title{4_Clonal_exploration}
\usage{
draw_violinplots(
  seq_df,
  group = "Patient_Sample",
  selected_rows,
  clonotype,
  AA_or_NT = "AA",
  region = "CDR3",
  percentage = 100,
  freq_filter = 0,
  Selected_clones = NULL,
  dominance_threshold,
  seed = 1234,
  really_hide_dots = F,
  width = 1400,
  height = 1000,
  img_type = "png",
  scale = 4,
  source = "clone_violinplot"
)
}
\value{
The return value, if any, from executing the utility.
}
\description{
A utils function
}
\keyword{internal}
