% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_2_Feature_determination.R
\name{Feature_1}
\alias{Feature_1}
\title{2_Feature_determination_1}
\usage{
Feature_1(path_base, grouping_by)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
\keyword{internal}
