% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_2_Feature_determination.R
\name{Feature__dataset}
\alias{Feature__dataset}
\title{2_Feature_determination_2}
\usage{
Feature__dataset(
  path_base,
  DF_to_parse,
  name_DF_to_parse,
  FWR1partial,
  FWR4partial,
  standard = T
)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
\keyword{internal}
