% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_3_Data_processing.R
\name{merge_FBMs}
\alias{merge_FBMs}
\title{Merge FBMs and meta files}
\usage{
merge_FBMs(filepath)
}
\arguments{
\item{filepath}{Path to the individual FBMs.}
}
\value{
A list with the merged FBM, the merged info file and the FBM header.
}
\description{
A function to join by row individual FBMs and .info files
    into a unique FBM and a unique .info file, merged_bm.
}
\keyword{internal}
