% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_1_Parsing.R
\name{parse_AIRRSeq_file}
\alias{parse_AIRRSeq_file}
\title{1_Parsing}
\usage{
parse_AIRRSeq_file(
  file,
  group,
  patient,
  subgroup,
  sample,
  input_path,
  C_region_included,
  FWR1partial,
  FWR4partial,
  D_gene,
  repertoire,
  output_path,
  is_example = F
)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
\keyword{internal}
